fix(1165): execute preflight checks before server startup
